% Generated by roxygen2 (4.0.1): do not edit by hand
\name{permF.mp}
\alias{permF.mp}
\title{Permutation F-tests for massively parallel linear models}
\usage{
permF.mp(formula, nperm = 499, alpha = 0.05, report.every = 50)
}
\arguments{
\item{formula}{a formula such as "\code{Y ~ X}", where \code{Y} is an \eqn{n
\times V} response matrix and \code{X} is an \eqn{n \times p} design matrix
common to all \eqn{V} models.}

\item{nperm}{number of permutations.}

\item{alpha}{level at which to control the FWER.}

\item{report.every}{parameter controlling how often to report the number of
permutations performed; by default, every \code{50}.}
}
\value{
\item{maxF.perm}{maximal F-statistics obtained from each of the
permuted data sets.} \item{F.obs}{the observed F-statistics.}
\item{threshold}{critical value obtained from the permutations.}
\item{pvalue}{adjusted (familywise error rate-controlling) p-values.}
}
\description{
Performs permutation F-tests for parallel linear models with a common design
matrix.  Currently restricted to testing with the intercept-only model as
the null hypothesis.  The permutation method controls the familywise error
rate (FWER) at a desired level; see Details.
}
\details{
The observed F-statistics are referred to a permutation distribution of the
maximum F-statistic over all \eqn{V} tests.  This is a standard approach to FWER
control in neuroimaging (Nichols and Holmes, 2001).
}
\examples{
Y = matrix(rnorm(6000), nrow=20)
X = rnorm(20)
t3 = permF.mp(Y~X)
}
\author{
Philip Reiss \email{phil.reiss@nyumc.org} and Lei Huang
\email{huangracer@gmail.com}
}
\references{
Nichols, T. E., and Holmes, A. P. (2001).  Nonparametric
permutation tests for functional neuroimaging: a primer with examples.
\emph{Human Brain Mapping}, 15, 1--25.
}
\seealso{
\code{\link{F.mp}}
}

