% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.rlrt4d}
\alias{plot.rlrt4d}
\title{Display cross-sections of voxelwise RLRT results}
\usage{
\method{plot}{rlrt4d}(x, array4d, disp = c("stat", "p", "fdr", "pwdf"),
  titl = NULL, slices = NULL, colbar = TRUE,
  col.image = shape::femmecol(100)[100:1], neglog10 = FALSE,
  threshold = NULL, mar = c(2, 2, 2, 2), digit = 2, nrow = NULL, ...)
}
\arguments{
\item{x}{a voxelwise RLRT object as produced by \code{\link{rlrt4d}}.}

\item{array4d}{the 4D array on which the voxelwise RLRT was performed.}

\item{disp,titl,neglog10,threshold}{see \code{\link{rlrtpanel}}.}

\item{slices}{indices of the slice(s) to be displayed.}

\item{colbar}{logical: Should a color bar be included?}

\item{col.image}{color scheme for the color bar, as generated by
\code{\link{rainbow}}, \code{\link{heat.colors}}, etc.}

\item{mar}{A numerical vector of the form c(bottom, left, top, right)
specifying the number of lines of margin on the four sides of the plot.}

\item{digit}{number of significant digits in labels.}

\item{nrow}{number of rows on the plot.}

\item{\dots}{arguments passed to \code{\link[graphics]{plot}.}}
}
\description{
Plots slices of the 3D array representing a set of voxelwise RLRT results.
}
\examples{
# Please see the example for rlrt4d
}
\author{
Lei Huang \email{huangracer@gmail.com}, Philip Reiss
\email{phil.reiss@nyumc.org} and Lan Huo
}
\seealso{
\code{\link{rlrt4d}}, \code{\link{rlrtpanel}}
}

