% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{test}
\alias{test}
\title{Toy data set}
\format{A list with two components: \describe{ \item{list("d4")}{a \eqn{10
\times 10 \times 10 \times 50} array of responses} \item{list("x")}{a vector
of 50 predictor values} }}
\description{
A randomly generated data set consisting of 50 "response" images and 50
scalar "predictors".
}
\keyword{datasets}

