\name{vows-internal}
\alias{get.ind}
\alias{rlr.xz}
\alias{rlrt.fit}
\alias{vec2img}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal functions for the vows package
}
\description{
These functions are ordinarily not to be called by the user, but if you contact the authors with any questions about them, we'll do our best to clarify matters.}
\author{
Lei Huang \email{huangracer@gmail.com}, Yin-Hsiu Chen \email{enjoychen0701@gmail.com}, and
Philip Reiss \email{phil.reiss@nyumc.org}
}


