% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R2nii.R
\name{R2nii}
\alias{R2nii}
\title{Save data to a NIfTI file}
\usage{
R2nii(arr, name.nii)
}
\arguments{
\item{arr}{a 3D or 4D array containing data to be saved.}

\item{name.nii}{filename, excluding the .nii extension.}
}
\value{
None; a NIfTI file is created.
}
\description{
This function can be used to output the results of voxelwise RLRT or
smoothing.
}
\author{
Lei Huang \email{huangracer@gmail.com}, Philip Reiss
\email{phil.reiss@nyumc.org} and Rong Jiao \email{jiaorong007@gmail.com}
}
\seealso{
\code{\link{nii2R}}
}

