\name{test}
\alias{test}
\docType{data}
\title{
  Toy data set
}
\description{
  A randomly generated data set consisting of 50 "response" images and 50 scalar "predictors".
}
\usage{data(test)}
\format{
A list with two components:
\describe{
\item{\code{d4}}{a \eqn{10 \times 10 \times 10 \times 50} array of responses}
\item{\code{x}}{a vector of 50 predictor values}
}
}
\keyword{datasets}
