% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vows-internal}
\alias{get.ind}
\alias{rlr.xz}
\alias{vec2img}
\alias{vows-internal}
\title{Internal functions for the vows package}

\description{
These functions are ordinarily not to be called by the user, but if you
contact the authors with any questions about them, we'll do our best to
clarify matters.
}
\author{
Lei Huang, Yin-Hsiu Chen, Philip Reiss \email{phil.reiss@nyumc.org} and Ruixin Tan
}

