% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vows-internal.R
\name{vows-internal}
\alias{get.ind}
\alias{rlr.xz}
\alias{vec2img}
\alias{vows-internal}
\title{Internal functions for the vows package}
\usage{
get.ind(arr3d)

rlr.xz(x, nbasis = 15, norder = 4, nulldim = NULL, B = NULL, P)

vec2img(vec, logicarray)
}
\description{
These functions are ordinarily not to be called by the user, but if you
contact the authors with any questions about them, we'll do our best to
clarify matters.
}
\author{
Lei Huang, Yin-Hsiu Chen, Philip Reiss \email{phil.reiss@nyumc.org} and Ruixin Tan
}
\keyword{internal}

